+++ python/cuml/ensemble/randomforestregressor.pyx
@@ -301,18 +301,18 @@ class RandomForestRegressor(Base):
         state = self.__dict__.copy()
         del state['handle']
         cdef size_t params_t
-        cdef  RandomForestMetaData[float, int] *rf_forest
-        cdef  RandomForestMetaData[double, int] *rf_forest64
+        cdef  RandomForestMetaData[float, float] *rf_forest
+        cdef  RandomForestMetaData[double, double] *rf_forest64
         cdef size_t params_t64
         if self.n_cols:
             # only if model has been fit previously
             model_pbuf_bytes = self._get_protobuf_bytes()
             params_t = <uintptr_t> self.rf_forest
             rf_forest = \
-                <RandomForestMetaData[float, int]*>params_t
+                <RandomForestMetaData[float, float]*>params_t
             params_t64 = <uintptr_t> self.rf_forest64
             rf_forest64 = \
-                <RandomForestMetaData[double, int]*>params_t64
+                <RandomForestMetaData[double, double]*>params_t64
             state["rf_params"] = rf_forest.rf_params
             state["rf_params64"] = rf_forest64.rf_params
         else:
@@ -355,9 +355,9 @@ class RandomForestRegressor(Base):
         # Only if model is fitted before
         # Clears the data of the forest to prepare for next fit
         if self.n_cols:
-            delete_rf_metadata(<RandomForestMetaData[float, int]*><uintptr_t>
+            delete_rf_metadata(<RandomForestMetaData[float, float]*><uintptr_t>
                                self.rf_forest)
-            delete_rf_metadata(<RandomForestMetaData[float, int]*><uintptr_t>
+            delete_rf_metadata(<RandomForestMetaData[double, double]*><uintptr_t>
                                self.rf_forest64)
 
             self.treelite_handle = None
